requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic>=2.5.0
pydantic-settings>=2.5.0
python-dotenv>=0.21.0
redis==5.0.1
openai==1.3.5
chromadb==0.4.18
pytest==7.4.3
pytest-asyncio==0.21.1
numpy==1.26.4
aiohttp==3.9.1
fhir.resources==7.1.0
